// frontend/src/api/generated/zod/recurring-transactions/recurring-transactions.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import { z as zod } from "zod";

export const getRecurringTransactionByIdParams = zod.object({
  id: zod.string().uuid(),
});

export const updateRecurringTransactionParams = zod.object({
  id: zod.string().uuid(),
});

export const updateRecurringTransactionBodyAmountMin = 0.01;

export const updateRecurringTransactionBody = zod.object({
  name: zod.string(),
  amount: zod.number().min(updateRecurringTransactionBodyAmountMin),
  description: zod.string().optional(),
  startDate: zod.string().date(),
  endDate: zod.string().date().optional(),
  scheduleType: zod.string(),
  categoryId: zod.string().uuid(),
});

export const deleteRecurringTransactionParams = zod.object({
  id: zod.string().uuid(),
});

export const createRecurringTransactionBodyAmountMin = 0.01;

export const createRecurringTransactionBody = zod.object({
  name: zod.string(),
  amount: zod.number().min(createRecurringTransactionBodyAmountMin),
  description: zod.string().optional(),
  startDate: zod.string().date(),
  endDate: zod.string().date().optional(),
  scheduleType: zod.string(),
  categoryId: zod.string().uuid(),
});

export const updateStatusParams = zod.object({
  id: zod.string().uuid(),
});

export const updateStatusBody = zod.object({
  isActive: zod.boolean(),
});